Phase 6 완료: 5대 혁신 기능을 main.py로 통합, 레거시 파일 정리
