--- s_class_dms_v19_main.py
@@ -1,597 +1,0 @@
-"""
-S-Class DMS v19.0: The Next Chapter - 메인 통합 시스템
-5대 혁신 기능을 통합한 차세대 지능형 운전자 모니터링 시스템
-"""
-
-import asyncio
-import time
-import logging
-from typing import Dict, List, Optional, Any
-from dataclasses import dataclass
-from pathlib import Path
-import json
-
-# Core Systems
-from config.settings import get_config, FeatureFlagConfig
-from models.data_structures import UIState
-from io_handler.ui import UIHandler
-
-# Innovation Systems (5대 혁신 기능)
-from systems.ai_driving_coach import AIDrivingCoach
-from systems.v2d_healthcare import V2DHealthcareSystem
-from systems.ar_hud_system import ARHUDSystem, VehicleContext
-from systems.emotional_care_system import EmotionalCareSystem
-from systems.digital_twin_platform import DigitalTwinPlatform
-
-
-@dataclass
-class SystemStatus:
-    """시스템 상태"""
-    ai_coach_active: bool = False
-    healthcare_active: bool = False
-    ar_hud_active: bool = False
-    emotional_care_active: bool = False
-    digital_twin_active: bool = False
-    
-    current_sessions: Dict[str, str] = None
-    last_update: float = 0.0
-    
-    def __post_init__(self):
-        if self.current_sessions is None:
-            self.current_sessions = {}
-
-
-class SClassDMSv19:
-    """S-Class DMS v19.0 메인 시스템"""
-    
-    def __init__(self, user_id: str = "default", edition: str = "RESEARCH"):
-        """
-        S-Class DMS v19.0 초기화
-        
-        Args:
-            user_id: 사용자 ID
-            edition: 에디션 (COMMUNITY, PRO, ENTERPRISE, RESEARCH)
-        """
-        self.config = get_config()
-        self.user_id = user_id
-        self.edition = edition
-        
-        # 피처 플래그 설정
-        self.feature_flags = FeatureFlagConfig(edition=edition)
-        
-        # 로깅 설정
-        self.logger = self._setup_logging()
-        
-        # UI 핸들러
-        self.ui_handler = UIHandler()
-        
-        # 시스템 상태
-        self.status = SystemStatus()
-        self.is_running = False
-        
-        # 혁신 시스템들 초기화
-        self.innovation_systems = self._initialize_innovation_systems()
-        
-        # 통합 데이터 저장
-        self.session_data = []
-        self.performance_metrics = {}
-        
-        # 동시 실행 태스크
-        self.running_tasks = []
-        
-        self.logger.info(f"S-Class DMS v19.0 시스템 초기화 완료")
-        self.logger.info(f"사용자: {user_id}, 에디션: {edition}")
-        self.logger.info(f"활성화된 기능: {self._get_enabled_features()}")
-
-    def _setup_logging(self) -> logging.Logger:
-        """로깅 설정"""
-        logger = logging.getLogger(f"SClassDMS_v19_{self.user_id}")
-        logger.setLevel(logging.INFO)
-        
-        # 핸들러가 이미 있으면 추가하지 않음
-        if not logger.handlers:
-            handler = logging.StreamHandler()
-            formatter = logging.Formatter(
-                '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
-            )
-            handler.setFormatter(formatter)
-            logger.addHandler(handler)
-        
-        return logger
-
-    def _initialize_innovation_systems(self) -> Dict[str, Any]:
-        """혁신 시스템들 초기화"""
-        systems = {}
-        
-        # 1. AI 드라이빙 코치 (PRO 이상)
-        if self.feature_flags.s_class_advanced_features:
-            try:
-                systems["ai_coach"] = AIDrivingCoach(self.user_id)
-                self.logger.info("✅ AI 드라이빙 코치 시스템 초기화 완료")
-            except Exception as e:
-                self.logger.error(f"❌ AI 드라이빙 코치 초기화 실패: {e}")
-        
-        # 2. V2D 헬스케어 플랫폼 (PRO 이상)
-        if self.feature_flags.s_class_advanced_features:
-            try:
-                systems["healthcare"] = V2DHealthcareSystem(self.user_id)
-                self.logger.info("✅ V2D 헬스케어 시스템 초기화 완료")
-            except Exception as e:
-                self.logger.error(f"❌ V2D 헬스케어 초기화 실패: {e}")
-        
-        # 3. AR HUD 시스템 (ENTERPRISE 이상)
-        if self.feature_flags.neural_ai_features:
-            try:
-                systems["ar_hud"] = ARHUDSystem()
-                self.logger.info("✅ AR HUD 시스템 초기화 완료")
-            except Exception as e:
-                self.logger.error(f"❌ AR HUD 초기화 실패: {e}")
-        
-        # 4. 감성 케어 시스템 (ENTERPRISE 이상)
-        if self.feature_flags.neural_ai_features:
-            try:
-                systems["emotional_care"] = EmotionalCareSystem(self.user_id)
-                self.logger.info("✅ 감성 케어 시스템 초기화 완료")
-            except Exception as e:
-                self.logger.error(f"❌ 감성 케어 초기화 실패: {e}")
-        
-        # 5. 디지털 트윈 플랫폼 (RESEARCH 에디션)
-        if self.feature_flags.innovation_research_features:
-            try:
-                systems["digital_twin"] = DigitalTwinPlatform()
-                self.logger.info("✅ 디지털 트윈 플랫폼 초기화 완료")
-            except Exception as e:
-                self.logger.error(f"❌ 디지털 트윈 플랫폼 초기화 실패: {e}")
-        
-        return systems
-
-    def _get_enabled_features(self) -> List[str]:
-        """활성화된 기능 목록"""
-        features = []
-        
-        if self.feature_flags.basic_expert_systems:
-            features.append("Expert Systems")
-        if self.feature_flags.s_class_advanced_features:
-            features.extend(["AI Coach", "Healthcare"])
-        if self.feature_flags.neural_ai_features:
-            features.extend(["AR HUD", "Emotional Care"])
-        if self.feature_flags.innovation_research_features:
-            features.append("Digital Twin")
-        
-        return features
-
-    async def start_system(self) -> bool:
-        """시스템 시작"""
-        try:
-            self.logger.info("🚀 S-Class DMS v19.0 시스템 시작")
-            
-            # UI 핸들러 시작
-            await self.ui_handler.start()
-            
-            # 각 혁신 시스템 세션 시작
-            await self._start_innovation_sessions()
-            
-            self.is_running = True
-            self.status.last_update = time.time()
-            
-            self.logger.info("✅ 모든 시스템이 정상적으로 시작되었습니다")
-            return True
-            
-        except Exception as e:
-            self.logger.error(f"❌ 시스템 시작 실패: {e}")
-            return False
-
-    async def _start_innovation_sessions(self):
-        """혁신 시스템 세션들 시작"""
-        
-        # AI 드라이빙 코치 세션
-        if "ai_coach" in self.innovation_systems:
-            try:
-                session_id = await self.innovation_systems["ai_coach"].start_driving_session()
-                self.status.current_sessions["ai_coach"] = session_id
-                self.status.ai_coach_active = True
-                self.logger.info(f"🎓 AI 코치 세션 시작: {session_id}")
-            except Exception as e:
-                self.logger.error(f"AI 코치 세션 시작 실패: {e}")
-        
-        # 헬스케어 세션
-        if "healthcare" in self.innovation_systems:
-            try:
-                session_id = await self.innovation_systems["healthcare"].start_health_monitoring()
-                self.status.current_sessions["healthcare"] = session_id
-                self.status.healthcare_active = True
-                self.logger.info(f"🏥 헬스케어 세션 시작: {session_id}")
-            except Exception as e:
-                self.logger.error(f"헬스케어 세션 시작 실패: {e}")
-        
-        # AR HUD 활성화
-        if "ar_hud" in self.innovation_systems:
-            self.status.ar_hud_active = True
-            self.logger.info("🥽 AR HUD 시스템 활성화")
-        
-        # 감성 케어 활성화
-        if "emotional_care" in self.innovation_systems:
-            self.status.emotional_care_active = True
-            self.logger.info("🎭 감성 케어 시스템 활성화")
-        
-        # 디지털 트윈 활성화
-        if "digital_twin" in self.innovation_systems:
-            self.status.digital_twin_active = True
-            self.logger.info("🤖 디지털 트윈 플랫폼 활성화")
-
-    async def run_main_loop(self):
-        """메인 실행 루프"""
-        self.logger.info("🔄 메인 실행 루프 시작")
-        
-        frame_count = 0
-        last_stats_time = time.time()
-        
-        try:
-            while self.is_running:
-                loop_start = time.time()
-                
-                # UI 상태 업데이트
-                ui_state = await self.ui_handler.update()
-                
-                if ui_state:
-                    # 모든 혁신 시스템에 데이터 전달 및 처리
-                    await self._process_with_innovation_systems(ui_state)
-                    
-                    # 세션 데이터 저장
-                    self.session_data.append({
-                        "timestamp": time.time(),
-                        "ui_state": ui_state,
-                        "frame_count": frame_count
-                    })
-                    
-                    # 성능 통계 (1초마다)
-                    if time.time() - last_stats_time >= 1.0:
-                        await self._update_performance_metrics()
-                        last_stats_time = time.time()
-                
-                # 프레임 레이트 제어 (30 FPS)
-                frame_time = time.time() - loop_start
-                sleep_time = max(0, 1/30 - frame_time)
-                await asyncio.sleep(sleep_time)
-                
-                frame_count += 1
-                
-                # 상태 업데이트
-                self.status.last_update = time.time()
-                
-        except KeyboardInterrupt:
-            self.logger.info("사용자에 의한 시스템 중단")
-        except Exception as e:
-            self.logger.error(f"메인 루프 오류: {e}")
-        finally:
-            await self.stop_system()
-
-    async def _process_with_innovation_systems(self, ui_state: UIState):
-        """혁신 시스템들과 함께 데이터 처리"""
-        
-        # 병렬 처리를 위한 태스크 리스트
-        tasks = []
-        
-        # 1. AI 드라이빙 코치 실시간 처리
-        if self.status.ai_coach_active and "ai_coach" in self.innovation_systems:
-            task = asyncio.create_task(
-                self.innovation_systems["ai_coach"].process_real_time_data(ui_state)
-            )
-            tasks.append(("ai_coach", task))
-        
-        # 2. 헬스케어 생체 데이터 처리
-        if self.status.healthcare_active and "healthcare" in self.innovation_systems:
-            task = asyncio.create_task(
-                self.innovation_systems["healthcare"].process_biometric_data(ui_state)
-            )
-            tasks.append(("healthcare", task))
-        
-        # 3. AR HUD 프레임 처리
-        if self.status.ar_hud_active and "ar_hud" in self.innovation_systems:
-            # 차량 컨텍스트 생성 (실제로는 차량 센서에서)
-            vehicle_context = VehicleContext(
-                speed_kmh=60.0,
-                steering_angle=0.0,
-                turn_signal=None,
-                gear="D"
-            )
-            
-            task = asyncio.create_task(
-                self.innovation_systems["ar_hud"].process_frame(ui_state, vehicle_context)
-            )
-            tasks.append(("ar_hud", task))
-        
-        # 4. 감성 케어 처리
-        if self.status.emotional_care_active and "emotional_care" in self.innovation_systems:
-            task = asyncio.create_task(
-                self.innovation_systems["emotional_care"].process_emotion_data(ui_state)
-            )
-            tasks.append(("emotional_care", task))
-        
-        # 모든 태스크 병렬 실행
-        if tasks:
-            try:
-                results = await asyncio.gather(*[task for _, task in tasks], return_exceptions=True)
-                
-                # 결과 처리
-                for i, (system_name, _) in enumerate(tasks):
-                    result = results[i]
-                    if isinstance(result, Exception):
-                        self.logger.error(f"{system_name} 처리 오류: {result}")
-                    else:
-                        await self._handle_system_result(system_name, result)
-                        
-            except Exception as e:
-                self.logger.error(f"병렬 처리 오류: {e}")
-
-    async def _handle_system_result(self, system_name: str, result: Any):
-        """시스템 결과 처리"""
-        
-        if system_name == "ai_coach" and result:
-            # AI 코치 피드백 처리
-            for feedback in result:
-                if feedback.priority <= 2:  # 높은 우선순위만 로그
-                    self.logger.info(f"🎓 AI 코치: {feedback.message}")
-        
-        elif system_name == "healthcare" and result:
-            # 헬스케어 경고 처리
-            for alert in result:
-                if alert.requires_medical_attention:
-                    self.logger.warning(f"🏥 건강 경고: {alert.message}")
-        
-        elif system_name == "ar_hud" and result is not None:
-            # AR HUD 프레임을 UI에 표시 (실제로는 HUD 디스플레이로 전송)
-            pass
-        
-        elif system_name == "emotional_care" and result:
-            # 감성 케어 세션 시작됨
-            self.logger.info(f"🎭 감성 케어 활성화: {result.care_mode.value}")
-
-    async def _update_performance_metrics(self):
-        """성능 메트릭 업데이트"""
-        try:
-            metrics = {
-                "timestamp": time.time(),
-                "session_data_count": len(self.session_data),
-                "active_systems": sum([
-                    self.status.ai_coach_active,
-                    self.status.healthcare_active,
-                    self.status.ar_hud_active,
-                    self.status.emotional_care_active,
-                    self.status.digital_twin_active
-                ])
-            }
-            
-            # 각 시스템별 통계
-            if "ai_coach" in self.innovation_systems:
-                coach_stats = self.innovation_systems["ai_coach"].get_driving_statistics()
-                metrics["ai_coach"] = coach_stats
-            
-            if "healthcare" in self.innovation_systems:
-                health_stats = self.innovation_systems["healthcare"].get_health_statistics()
-                metrics["healthcare"] = health_stats
-            
-            if "ar_hud" in self.innovation_systems:
-                ar_stats = self.innovation_systems["ar_hud"].get_ar_statistics()
-                metrics["ar_hud"] = ar_stats
-            
-            if "emotional_care" in self.innovation_systems:
-                care_stats = self.innovation_systems["emotional_care"].get_care_statistics()
-                metrics["emotional_care"] = care_stats
-            
-            if "digital_twin" in self.innovation_systems:
-                twin_stats = self.innovation_systems["digital_twin"].get_platform_statistics()
-                metrics["digital_twin"] = twin_stats
-            
-            self.performance_metrics = metrics
-            
-        except Exception as e:
-            self.logger.error(f"성능 메트릭 업데이트 오류: {e}")
-
-    async def stop_system(self):
-        """시스템 중단"""
-        self.logger.info("🛑 S-Class DMS v19.0 시스템 중단 시작")
-        
-        self.is_running = False
-        
-        # 혁신 시스템 세션들 종료
-        await self._stop_innovation_sessions()
-        
-        # UI 핸들러 중단
-        if hasattr(self, 'ui_handler'):
-            await self.ui_handler.stop()
-        
-        # 실행 중인 태스크들 정리
-        for task in self.running_tasks:
-            if not task.done():
-                task.cancel()
-        
-        # 최종 데이터 저장
-        await self._save_session_data()
-        
-        self.logger.info("✅ 모든 시스템이 정상적으로 종료되었습니다")
-
-    async def _stop_innovation_sessions(self):
-        """혁신 시스템 세션들 종료"""
-        
-        # AI 드라이빙 코치 세션 종료
-        if self.status.ai_coach_active and "ai_coach" in self.innovation_systems:
-            try:
-                result = await self.innovation_systems["ai_coach"].end_driving_session()
-                self.logger.info(f"🎓 AI 코치 세션 종료: 점수 {result.get('session_report', {}).get('overall_score', 0):.1f}")
-                self.status.ai_coach_active = False
-            except Exception as e:
-                self.logger.error(f"AI 코치 세션 종료 오류: {e}")
-        
-        # 헬스케어 세션 종료
-        if self.status.healthcare_active and "healthcare" in self.innovation_systems:
-            try:
-                result = await self.innovation_systems["healthcare"].end_health_session()
-                self.logger.info(f"🏥 헬스케어 세션 종료: {result.get('metrics_count', 0)}개 메트릭 수집")
-                self.status.healthcare_active = False
-            except Exception as e:
-                self.logger.error(f"헬스케어 세션 종료 오류: {e}")
-        
-        # 감성 케어 세션 종료 (활성 중인 경우)
-        if self.status.emotional_care_active and "emotional_care" in self.innovation_systems:
-            try:
-                if self.innovation_systems["emotional_care"].is_care_active:
-                    result = await self.innovation_systems["emotional_care"].end_care_session()
-                    self.logger.info(f"🎭 감성 케어 세션 종료: 효과성 {result.get('effectiveness_score', 0):.2f}")
-                self.status.emotional_care_active = False
-            except Exception as e:
-                self.logger.error(f"감성 케어 세션 종료 오류: {e}")
-
-    async def _save_session_data(self):
-        """세션 데이터 저장"""
-        try:
-            session_dir = Path("sessions") / f"session_{int(time.time())}"
-            session_dir.mkdir(parents=True, exist_ok=True)
-            
-            # 메타데이터 저장
-            metadata = {
-                "user_id": self.user_id,
-                "edition": self.edition,
-                "session_start": self.session_data[0]["timestamp"] if self.session_data else time.time(),
-                "session_end": time.time(),
-                "total_frames": len(self.session_data),
-                "active_systems": {
-                    "ai_coach": self.status.ai_coach_active,
-                    "healthcare": self.status.healthcare_active,
-                    "ar_hud": self.status.ar_hud_active,
-                    "emotional_care": self.status.emotional_care_active,
-                    "digital_twin": self.status.digital_twin_active
-                },
-                "performance_metrics": self.performance_metrics
-            }
-            
-            metadata_file = session_dir / "metadata.json"
-            with open(metadata_file, 'w', encoding='utf-8') as f:
-                json.dump(metadata, f, ensure_ascii=False, indent=2)
-            
-            self.logger.info(f"세션 데이터 저장 완료: {session_dir}")
-            
-        except Exception as e:
-            self.logger.error(f"세션 데이터 저장 오류: {e}")
-
-    async def create_digital_twin_from_session(self) -> Optional[str]:
-        """현재 세션 데이터로 디지털 트윈 생성"""
-        if not self.status.digital_twin_active or "digital_twin" not in self.innovation_systems:
-            self.logger.warning("디지털 트윈 플랫폼이 비활성화되어 있습니다")
-            return None
-        
-        if len(self.session_data) < 100:
-            self.logger.warning("디지털 트윈 생성을 위한 데이터가 부족합니다 (최소 100 프레임 필요)")
-            return None
-        
-        try:
-            # UI 상태 데이터 추출
-            ui_states = [data["ui_state"] for data in self.session_data]
-            session_ids = [f"session_{int(time.time())}"]
-            
-            # 디지털 트윈 생성
-            digital_twin = await self.innovation_systems["digital_twin"].create_digital_twin(
-                ui_states, session_ids
-            )
-            
-            self.logger.info(f"🤖 디지털 트윈 생성 완료: {digital_twin.twin_id}")
-            return digital_twin.twin_id
-            
-        except Exception as e:
-            self.logger.error(f"디지털 트윈 생성 오류: {e}")
-            return None
-
-    def get_system_status(self) -> Dict[str, Any]:
-        """시스템 상태 조회"""
-        return {
-            "is_running": self.is_running,
-            "user_id": self.user_id,
-            "edition": self.edition,
-            "enabled_features": self._get_enabled_features(),
-            "active_systems": {
-                "ai_coach": self.status.ai_coach_active,
-                "healthcare": self.status.healthcare_active,
-                "ar_hud": self.status.ar_hud_active,
-                "emotional_care": self.status.emotional_care_active,
-                "digital_twin": self.status.digital_twin_active
-            },
-            "current_sessions": self.status.current_sessions,
-            "session_frames": len(self.session_data),
-            "last_update": self.status.last_update,
-            "performance_metrics": self.performance_metrics
-        }
-
-    async def run_digital_twin_simulation(self, twin_id: str, scenario_count: int = 100) -> Dict[str, Any]:
-        """디지털 트윈 시뮬레이션 실행"""
-        if not self.status.digital_twin_active or "digital_twin" not in self.innovation_systems:
-            return {"error": "디지털 트윈 플랫폼이 비활성화되어 있습니다"}
-        
-        try:
-            platform = self.innovation_systems["digital_twin"]
-            
-            # 시나리오 생성
-            scenarios = await platform.generate_simulation_scenarios(count=scenario_count)
-            
-            # 시뮬레이션 실행
-            results = await platform.run_mass_simulation(twin_id, scenarios)
-            
-            # 결과 분석
-            analysis = await platform.analyze_simulation_data(results)
-            
-            # AI 모델 개선
-            improvements = await platform.improve_ai_models(results)
-            
-            return {
-                "simulation_results": len(results),
-                "success_rate": sum(1 for r in results if r.success) / len(results),
-                "analysis": analysis,
-                "model_improvements": improvements
-            }
-            
-        except Exception as e:
-            self.logger.error(f"디지털 트윈 시뮬레이션 오류: {e}")
-            return {"error": str(e)}
-
-
-async def main():
-    """메인 실행 함수"""
-    print("🚀 S-Class DMS v19.0: The Next Chapter")
-    print("=" * 50)
-    
-    # 시스템 초기화
-    dms_system = SClassDMSv19(
-        user_id="test_user", 
-        edition="RESEARCH"  # 모든 기능 활성화
-    )
-    
-    try:
-        # 시스템 시작
-        if await dms_system.start_system():
-            print("✅ 시스템이 성공적으로 시작되었습니다!")
-            print("\n현재 활성화된 혁신 기능들:")
-            
-            status = dms_system.get_system_status()
-            for system, active in status["active_systems"].items():
-                status_icon = "🟢" if active else "🔴"
-                print(f"  {status_icon} {system}")
-            
-            print("\n메인 실행 루프를 시작합니다...")
-            print("중단하려면 Ctrl+C를 누르세요.\n")
-            
-            # 메인 루프 실행
-            await dms_system.run_main_loop()
-            
-        else:
-            print("❌ 시스템 시작에 실패했습니다.")
-            
-    except KeyboardInterrupt:
-        print("\n사용자에 의한 시스템 중단")
-    except Exception as e:
-        print(f"❌ 오류 발생: {e}")
-    finally:
-        print("시스템을 정리하는 중...")
-
-
-if __name__ == "__main__":
-    # asyncio 실행
-    asyncio.run(main())
